Organize files into folders. Adjust namespaces
