+++ src/CatBox.NET/Enums/UploadHost.cs
@@ -1,4 +1,4 @@
-﻿namespace CatBox.NET;
+﻿namespace CatBox.NET.Enums;
 
 /// <summary>
 /// Currently unused
