+++ src/CatBox.NET/Requests/CreateAlbumRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace CatBox.NET;
+﻿namespace CatBox.NET.Requests;
 
 /// <summary>
 /// Wraps a request to create a new album and add files to it
