+++ src/CatBox.NET/Requests/DeleteFileRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace CatBox.NET;
+﻿namespace CatBox.NET.Requests;
 
 /// <summary>
 /// Wraps a request to delete files from the API
